requirements.txt
asgiref==3.7.2
Django==4.2.3
django-filter==23.2
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
drf-yasg==1.21.6
inflection==0.5.1
packaging==23.1
phonenumbers==8.13.15
Pillow==10.0.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
